Links to Others' Pages
In my "See Also" section, I've added a few extra links to other students' pages, since the rubric required that I have some of those. One of the student's pages (And this is a fact shown by the comments) just seems to be an empty page, though I suppose this may just be because they haven't pushed. Regardless, I put it there because the subject it's said to have on the spreadsheet (Jaguars) was the most relevant to my page of those I saw.