Finished Main Component
Finally finished the main component. The image file I'm using is JPEG and not JPG, so I had to change that in the html file. In the CSS file, my main problem was that, for the article element selector, I had used color instead of background-color... Moving on to making my custom card (Have also added the M logo image)