Fixed CSS + a few additions
Fixed the CSS and finally got the background image's scale working (I was writing "img #background" as my selector, completely forgetting that an ID is for a very specific element, and does not use those modifiers.) Also added a span selector to adjust the colors of the kanji. About to change a lot of stuff, including the background image.