Done. See desc.
Caught up with the work I lost, and as was the case with part 1, learned a lot about gaps in my knowledge in doing so. I've used a slightly janky method to get the grid to work on these pages, so maybe I'll change that if I have time tonight. Moving on to finishing part 3.