CSS File Path Change
Though the path I originally had was correct, it did not work on GitHub Pages, as I used a / at the start to access everything from the root. Locally, the root was MaxMosier, but on GutHub Pages, it seems it's maxmosier.github.io, so it was trying to access something like maxmosier.github.io/css/ica6/...

While I could have just replaced the first / with /MaxMosier/, this would have stopped things from working locally. So, I did everything relative, and used "../../" in place of the / . It should work now, but I'll have to wait for the page to update to know for sure.