Done (?)
As far as I can tell, I'm done with this assignment.  I made a lot of changes, all of which should be rather obvious, or if not, should have explaining comments, at the very least. Hardest part was finding a way to center the images when the vertical layout resulted in the content area of .grid being larger than the images themselves. When smaller, the width:auto would cause them to take up the whole width, centering automatically. When in this small range in which the content was bigger, images would default to the left side. Though I could probably have centered them all using other, simpler methods, this is the only way I've been able to devise, at the moment.