Lots of Changes
Changed the font, added a glow to the nav and h1, changed the font color and background color, added padding to nav a text, added glow animation to hover of anchors in nav, changed the title and h1 text, and the color of the ball. There's probably more, but anything I missed can just be seen in the changes.