Color Changes and Line Directionality
Added the ability to alter the directions of the warning line patterns inline! The color can also be changed inline. Note that the --warning-line-color property is no longer in html, but in the stripeBox selector. Also made it so that list items alternate in text and background color. Might make it a bit different later, but this is what I'm doing for now!