Motion, Background, Circle Mask
Created the ability to vary the stripes' speeds, and used said ability to slow the inner stripes more. Also tried changing the background to be a radial gradient because I wanted a vignette, but got this cool retro effect, apparently??  I have no idea what's causing it, but I wanted to push it to save it before I "fix" it. Also added a circular mask to the backgrounds of the nav bar's elements!