Menu Animation, Additional Background Variables
Added an animation to the menu item currently selected, which was a bit of a doozy. Had to double the length of each list item, and top it off (off screen) with a solid-colored gradient, since I can't animate background images, but CAN translate them. When the background image is shifted left by say 1em, it fills the empty space with the leftmost end of the image, in a tiling fashion. However, I previously had full transparency at the end. That should be fixed now!

Also added another defenition at the top for the shade of grey used by the background of the html element. It's just mathematically defined to be a ratio of the original grey, and at the moment, is half the brightness.