Rearrangement and Link Addition
Rearranged the CSS to have an order that makes a bit more sense, by placing the logo closer to the top. Sure, it appears later in the HTML, but it's more prominent on the page, and more related to the title, which is on top. The complicated stripe stuff should go on the bottom!

Also added the ability to actually click on the links, and attempted to make the entire block clickable. It works horizontally, but for whatever reason, the anchor isn't filling the entire container vertically... Oh well! It's better than what I had.

About to try and add one final thing to the dropdown, and then I'll get started on making this adaptive.