ID Fixes, Emboldening of Shadows
Realized after validating that I had many instances in which I was applying an ID to many different elements, where I should have been using a class. Converted all of said instances to classes.

Also changed the effect on the M image from a glow to a drop shadow, and because the drop-shadow filter hasn't a high opacity, I tripled it by just typing it three times.

I spent some time trying to make a special element which would dim all other menu options when they were selected, but I couldn't get that working, and figured the benefits of that were not worth the time cost for now. Perhaps I'll come back to that when we get into JavaScript, since I've already been quite restricted with what I can do in pure CSS...

Finally getting on to making the alternative layouts for other screens!