(Some) Responsiveness Added
Added a bit of responsiveness to the nav elements, and the image elements. There were a few distinct stages at which they broke down, so I dealt with those. There may be a few straggling breaking points I haven't observed yet, but those would require some really odd screens or layouts to be encountered, so I'm not gonna' deal with them unless I've more time to do so later. Pushing!