More Media Queries
Adding some more support for portrait orientations, as I have so far only dealt with the landscapes. I'm about to make a copy of this file called "index.css", as that's what is required I do based on the rubric. After that, I'm heading to bed, so I'll finish this up tomorrow! Gonna' turn this in right now, though.