Nav Bar Portrait + M Logo Troubles
Have given the nav bar a basic layout for portrait mode, though there are plenty of breakdown cases I'm yet to look in to.

Currently, the main problem is the M logo, which I for some reason defined the width of by the viewport height last night... Gonna' start fixing that before I deal with nav bar breakdowns, since it's really obvious and annoying.