Responsiveness Done
Okay, I think I'm done with responsiveness. Whether this is out of a desire to stop dealing with this or if I am truly done may be something I'm failing to distinguish, but as far as I can tell, the screen works for most ratios and sizes!

Next up: Create a new page for the "About me" content, and fill it with something. Then, come up with something to provide as the "Resume" link, and another to use in place of what is currently "OPTION @"

Once I've done all of that, I might be copying this whole thing to a new repository, as the assignment requires this page be the one accessed by maxmosier.github.io, which is currently not the case.