Scrollbar and Box Problems
Having a lot of trouble getting the text box to fit properly... It's doing some really odd things I don't fully understand. When its width is set to 10vw, it's showing up as what looks to be 20vw...

However, I did learn how to customize scrollbars, so I made one of those for the box!