Added ForestSize
Instead of using the weight, went ahead and used an area unit instead, since that was easier to insert into the story. Also note that on line 96, I changed "textContent" to "innerHTML" in order to get the exponent after the area units to display properly (Because "textContent" resulting in <sup>2</sup> being displayed as raw text).