Removal of Comment
Removed a comment I had placed because instead of copy-pasting from the given template on MDN, I had been typing everything out. The comment was there to remind me that I had to check this section for errors, since I couldn't quite see the text on the projector (I was doing this whole thing in class, but for some reason couldn't access the MDN from my own device, and so had read off of the screen).