Another Misspelled Variable
Apparently when I was typing what I saw on the projector in class (I didn't have access to MDN at the time), I typed "button" instead of "btn" when declaring the variable used to listen for clicks. That's fine, but I didn't change that name on the very next line, apparently. I changed it back to "btn," since the element type "button" already exists. everything seems to be working now!