Fixed Paths
I realized today that my README does not currently link to my CSS and HTML projects! In deciding which link to paste for this HTML midterm, I remembered that the feedback I got on the midterm was that none of the cross-page paths in the nav bar worked. Coming back to fix that so I can update my README (And of course, for the sake of making it right!)