Added CSS and HTML Midterm Projects
As the title implies, I've added the two projects I've so far finished, after realizing I never updated the README for those a few minutes ago. As a side note, I also just finished fixing a part of my HTML midterm in order to get it working on this page.