Image Uploads
Just uploaded a bunch of random images from my computer for use in this project, but just realized I should have resized them to be 640 by 480, as opposed to the 640 by 640 that I actually did. Gotta' find and make a new set of images, but I'll keep these here for later, in case I need some small square placeholder images!