Functional
Now that I'm on my laptop, I was better able to see all of the errors I had so far created. They were really just misnaming errors, since the original template for the assignment had been based on quotes, while the new version of the assignment regards questions (Instead of quotes).

About to add the ability to write an answer, but am pushing because I plan to make a pretty significant change.