Refresh Button Progress
Beginning to add functionality for the refresh button, but am having a single problem with getting it to work: "Uncaught TypeError: Cannot read properties of null (reading 'addEventListener')".

The only thing I can think of is that the page is loading the button AFTER the script, somehow. I don't really see why that would happen, but I'm gonna' try and force it to only start the script once all elements are loaded with DOM or something.