Image Change
The image I was using looked really bad as a background, so I've changed it to this new one! It doesn't really show the butte, which is actually super unfortunate, but due to the layout of my page, I would have had to find one where the butte was in to the side.