Japanese
Added a function which converts all of the date names to the corresponding kanji used for the days of the week in Japan. This was done to make things a bit cleaner, and to make the width of the divs more consistent (In other words, was a slightly lazy way of making it so I don't have to mess with the divs too much.