Fixing Days of Week
Apparently the weather.gov api will also include "this afternoon" if you view it early enough. As a result, the way I had prepared to receive the days of the week was incorrect, and it was messing everything up.

Just fixed that, and also (Though this doesn't actually do anything) added some code to the Japanese day conversion function which'll prepare it to receive those erroneous strings, just in case.