Getting Started
Getting started on this! So far, all I've done is add a loop which adds the names of all of the mother cats to the motherInfo string. Also worth noting (And I made a comment where I did this) is that I changed the original file's ".text" on line 39 to ".json," in order to load in the info as a json.

I also see that this calls for some stylesheet which is not actually given, so for the time being, I've commented that out.