Done!
Almost made a separate loop to count the kittens and whatnot, but realized I could fit everything in one loop, so that's what I did! As far as I can tell, it's correct and done.

Also note that I removed the styling block at the top and re-added the reference to the styling sheet the file originally had.