Misnaming + RGB->HSL
For what is FINALLY the last time, I've fixed the mispellings I made while viewing the love coding in class. We were told to just copy/paste the update code from a website in class, and though that code was perfectly valid, it used variables whom I had given different names. Fixed! This seems to be working now, and I can finally begin implementing a collision function.

* Meant to push this, but apparently didn't before making the following changes:

Changed the RGB function to HSV, in order to make it easier to control the colors.