File Path Changes
Before realizing I didn't yet have to write anything on the files for this assignment, I had been preparing to turn it in as "quiz13," under the impression that its being named "wa13" was a mistake. Had changed the file paths accordingly, but am now basically reverting them back to being what they once were, albeit in such a manner as to follow my file structure.