Prototype Finished
I didn't really make that many pushes, but have so far completed a basic, functional outline of what I want to do for this assignment. I'm gonna' change a few things up, but it works decently at the moment.

One change that is to come will include a complete refactoring of the JS. Instead of doing what it does now to determine the volume, which is to average the interval between the past X clicks, I need it to count the number of clicks over the past X (milli)seconds. Completely different concept!