Ignoring Copyrighted Files
I've been using the Don't Starve soundtrack to test my wa13 page locally, but don't want to upload the contents to GitHub, since I assume that's a form of redistribution, and assume that's not something I'm allowed to do. Adding those files to gitignore so I don't accidentally do that!