YouTube Version
Made a version of this WA which uses the YouTube API to source audio. This was done because I didn't want to upload the files to the music I had, since that would be redistribution of the files, which is something I don't want to be doing! It seems to work well! However, it does randomly seem to not play at times, and also doesn't work at all with some videos, according to my tests.