Volume Remapper
To make the increases/decreases in volume more noticeable, made it so that the volume is actually changed using an exponential curve. The difference between 10 and 20 was previously far more noticeable than that between 80 and 90, for example. Adding the exponential function makes it feel far more sensitive and even in ascent.