Removed Zooming
On Apple devices, tapping rapidly just caused the device to zoom in and out on the page, and made it entirely unusable. What I've just done SHOULD remove that, but I'll find out in about 10 mins, after the page updates. Zooming shouldn't be needed on the page, and if I think the text or button are too small, I'll just enlarge them in the CSS.