Upload
I've been doing this all locally, but am gonna' start working on it through my actual GitHub now. I'm currently in the process of finding a way to make this an actual webpage, since so far I've just been messing with the JS. Not sure what I want to put here, but I'll think of something! It's a bit noisy at the moment...