CSS reformatting + Temp effect
Made the HTML element a bit easier to adjust using variables. Also, just for fun, made another div which has opaque stripes overlaid on the entities, so it looks like they're underneath it! That's just for testing, though. I'm gonna' remove it after this push, but just wanted to record it!