Scrollbar and Fixes
Got a custom scrollbar set up.
As of the last push, it was impossible to interact with the sliders and buttons, and I discovered that this was because it was for some reason being placed above everything. So, I have it a z-index of -1, and that fixed the problem! It seems to be working!