Adding, Subtracting, Range Altering, Adjusting
Added functionality to my second slider and the three buttons!

The randomize function is pretty obvious. Just moves and rotates all of the present entities randomly (and does NOT change their speeds).

The add and subtract entity functions do just as they say, with a minimum limit of 10 entities and a subjectively set maximum of 750. Also changed the default to 200, since you can now manage them yourself. I figured 10 seemed like a good increment.

And the slider is pretty simple, too. Just changes the aRange and aRangeMin variables! aRangeMin is bound to a range of 0 to 45 (subjectively decided on) but within that range is always 3/5ths of the aRange.

Next up, gonna make the entity transparency dependent on its speed.