!POST-SUBMISSION CHANGE! Quick Button Fix
!!!POST-SUBMISSION CHANGE!!!

After trying this on a device with a higher pixel density than what I've been using, I found that my buttons were set up in a bit of an odd fashion! Fixed that. This IS a commit I'm making post-submission, though, so disregard its effects. It's just for my own sanity that I made this change.